+++ src/component/base/methods.js
@@ -24,6 +24,9 @@ import { removeGlobalEffects } from '../../lib/reactivity/effect.js'
 
 export default {
   focus: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (e) {
       Log.warn('this.focus is deprecated, use this.$focus instead')
       return this.$focus(e)
@@ -33,7 +36,11 @@ export default {
     configurable: false,
   },
   $focus: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (e) {
+      if (this.eol === true) return
       this[symbols.state].hasFocus = true
       Focus.set(this, e)
     },
@@ -42,6 +49,9 @@ export default {
     configurable: false,
   },
   unfocus: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function () {
       this[symbols.state].hasFocus = false
       this.lifecycle.state = 'unfocus'
@@ -51,16 +61,58 @@ export default {
     configurable: false,
   },
   destroy: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function () {
+      this.eol = true
       this.lifecycle.state = 'destroy'
+
+      for (let key in this[symbols.state]) {
+        if (Array.isArray(this[symbols.state][key])) {
+          this[symbols.state][key] = []
+        }
+      }
+
       this.$clearTimeouts()
       this.$clearIntervals()
       eventListeners.removeListeners(this)
       deleteChildren(this[symbols.children])
+      this[symbols.children].length = 0
       removeGlobalEffects(this[symbols.effects])
+
       this[symbols.state] = {}
+
       this[symbols.props] = {}
+      this[symbols.computed] = null
+      this.lifecycle = {}
       this[symbols.effects].length = 0
+      this.parent = null
+      this.rootParent = null
+      this[symbols.wrapper] = null
+      this[symbols.originalState] = null
+      this[symbols.slots].length = 0
+
+      delete this[symbols.computed]
+      delete this[symbols.effects]
+      delete this.parent
+      delete this.rootParent
+      delete this[symbols.wrapper]
+      delete this[symbols.originalState]
+      delete this[symbols.children]
+      delete this[symbols.slots]
+      delete this.componentId
+      delete this[symbols.id]
+      delete this.ref
+      delete this[symbols.state].hasFocus
+
+      this[symbols.holder].destroy()
+      this[symbols.holder] = null
+      delete this[symbols.holder]
+
+      this[symbols.cleanup]()
+      delete this[symbols.cleanup]
+
       Log.debug(`Destroyed component ${this.componentId}`)
     },
     writable: false,
@@ -68,6 +120,9 @@ export default {
     configurable: false,
   },
   [symbols.removeGlobalEffects]: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (effects = []) {
       removeGlobalEffects(effects)
     },
@@ -76,6 +131,9 @@ export default {
     configurable: false,
   },
   select: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (ref) {
       Log.warn('this.select is deprecated, use this.$select instead')
       return this.$select(ref)
@@ -85,6 +143,9 @@ export default {
     configurable: false,
   },
   $select: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (ref) {
       let selected = null
       this[symbols.children].forEach((child) => {
@@ -107,6 +168,9 @@ export default {
     configurable: false,
   },
   trigger: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (key) {
       Log.warn('this.trigger is deprecated, use this.$trigger instead')
       return this.$trigger(key)
@@ -116,6 +180,9 @@ export default {
     configurable: false,
   },
   $trigger: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (key) {
       let target = this[symbols.originalState]
       // when dot notation used, find the nested target
@@ -134,6 +201,9 @@ export default {
     configurable: false,
   },
   shader: {
+    /**
+     * @this {import('../../component').BlitsComponent}
+     */
     value: function (type, args) {
       return {
         type: type,
@@ -156,6 +226,10 @@ export default {
   },
 }
 
+/**
+ * Recursively destroys all BlitsComponent children.
+ * @param {import('../../component').BlitsComponent[]} children
+ */
 const deleteChildren = function (children) {
   for (let i = 0; i < children.length; i++) {
     if (!children[i]) return